Separate some extensions and remove unused var
